+++ src/event/internal.rs
@@ -3,6 +3,8 @@ use std::time::Duration;
 use parking_lot::{MappedMutexGuard, Mutex, MutexGuard};
 
 #[cfg(unix)]
+use crate::colors::{ColorEntry, ColorScheme};
+#[cfg(unix)]
 use crate::event::KeyboardEnhancementFlags;
 use crate::event::{filter::Filter, read::InternalEventReader, timeout::PollTimeout, Event};
 
@@ -59,77 +61,6 @@ where
 {
     let mut reader = lock_event_reader();
     reader.try_read(filter)
-}
-
-/// Terminal color type, used in queries and responses.
-///
-/// `Palette(n)` uses OSC 4. The remaining variants use OSC 10..=19.
-#[derive(Debug, PartialOrd, PartialEq, Hash, Clone, Copy, Eq)]
-pub enum ColorType {
-    Palette(u8),
-    Foreground,
-    Background,
-    Cursor,
-    PointerForeground,
-    PointerBackground,
-    TektronixForeground,
-    TektronixBackground,
-    HighlightBackground,
-    TektronixCursor,
-    HighlightForeground,
-}
-
-impl ColorType {
-    /// Maps an OSC number (10..=19) to the corresponding `ColorType` variant.
-    pub(crate) fn from_osc_number(n: u8) -> Option<Self> {
-        match n {
-            10 => Some(Self::Foreground),
-            11 => Some(Self::Background),
-            12 => Some(Self::Cursor),
-            13 => Some(Self::PointerForeground),
-            14 => Some(Self::PointerBackground),
-            15 => Some(Self::TektronixForeground),
-            16 => Some(Self::TektronixBackground),
-            17 => Some(Self::HighlightBackground),
-            18 => Some(Self::TektronixCursor),
-            19 => Some(Self::HighlightForeground),
-            _ => None,
-        }
-    }
-
-    /// Returns the OSC number for this color type.
-    pub(crate) fn osc_number(&self) -> u8 {
-        match self {
-            Self::Palette(_) => 4,
-            Self::Foreground => 10,
-            Self::Background => 11,
-            Self::Cursor => 12,
-            Self::PointerForeground => 13,
-            Self::PointerBackground => 14,
-            Self::TektronixForeground => 15,
-            Self::TektronixBackground => 16,
-            Self::HighlightBackground => 17,
-            Self::TektronixCursor => 18,
-            Self::HighlightForeground => 19,
-        }
-    }
-}
-
-/// The terminal's color scheme preference (dark or light).
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug, PartialOrd, Ord, PartialEq, Hash, Clone, Copy, Eq)]
-pub enum ColorScheme {
-    Dark,
-    Light,
-}
-
-/// A parsed color response from the terminal.
-#[derive(Debug, PartialOrd, PartialEq, Hash, Clone, Eq)]
-pub(crate) struct ColorEntry {
-    pub color_type: ColorType,
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
 }
 
 /// An internal event.
